Enforce use of random port numbers for local endpoints.
This might seem like a limitation, but the more I think about it,
the more I am convinced that this is the right decision and that
the restrictions it puts in place will eventually enable quite
interesting backend architectures.

While there, add "zeromq_single" directive to disable this
enforcement (only for testing and development).

Change-Id: I6cb72c227c8581a907047b1c70edd2d5774cfbdd
Signed-off-by: Piotr Sikora <piotr.sikora@frickle.com>